Removing the old sprite editor (may it rest in peace)
