+++ webapp/src/components/ImageEditor/ImageCanvas.tsx
@@ -3,13 +3,12 @@ import { connect } from 'react-redux';
 
 import { ImageEditorStore, ImageEditorTool } from './store/imageReducer';
 import { dispatchImageEdit, dispatchChangeZoom, dispatchChangeCursorLocation } from "./actions/dispatch";
-import { ImageState, Bitmap } from './store/bitmap';
 import { GestureTarget, ClientCoordinates, bindGestureEvents } from './util';
 
 import { Edit, EditState, getEdit, getEditState, ToolCursor, tools } from './toolDefinitions';
 
 export interface ImageCanvasProps {
-    dispatchImageEdit: (state: ImageState) => void;
+    dispatchImageEdit: (state: pxt.sprite.ImageState) => void;
     dispatchChangeZoom: (zoom: number) => void;
     dispatchChangeCursorLocation: (loc: [number, number]) => void;
     selectedColor: number;
@@ -20,8 +19,8 @@ export interface ImageCanvasProps {
     onionSkinEnabled: boolean;
 
     colors: string[];
-    imageState: ImageState;
-    prevFrame?: ImageState;
+    imageState: pxt.sprite.ImageState;
+    prevFrame?: pxt.sprite.ImageState;
 }
 
 /**
@@ -370,7 +369,7 @@ class ImageCanvasImpl extends React.Component<ImageCanvasProps, {}> implements G
         }
     }
 
-    protected drawBitmap(bitmap: Bitmap, x0 = 0, y0 = 0, transparent = true) {
+    protected drawBitmap(bitmap: pxt.sprite.Bitmap, x0 = 0, y0 = 0, transparent = true) {
         const { colors } = this.props;
 
         const context = this.canvas.getContext("2d");
