+++ webapp/src/components/ImageEditor/store/imageReducer.ts
@@ -1,5 +1,4 @@
 import * as actions from '../actions/types'
-import { ImageState, Bitmap } from './bitmap';
 
 export enum ImageEditorTool {
     Paint,
@@ -32,7 +31,7 @@ export interface AnimationState {
     aspectRatioLocked: boolean;
 
     currentFrame: number;
-    frames: ImageState[];
+    frames: pxt.sprite.ImageState[];
     interval: number;
 }
 
@@ -204,7 +203,7 @@ const animationReducer = (state: AnimationState, action: any): AnimationState =>
                 ...state,
                 frames: state.frames.map((frame, index) => ({
                     ...frame,
-                    bitmap: Bitmap.fromData(frame.bitmap).resize(width, height).data()
+                    bitmap: pxt.sprite.Bitmap.fromData(frame.bitmap).resize(width, height).data()
                 }))
             };
         case actions.IMAGE_EDIT:
@@ -304,17 +303,17 @@ const editorReducer = (state: EditorState, action: any): EditorState => {
     return state;
 }
 
-function emptyFrame(width: number, height: number): ImageState {
+function emptyFrame(width: number, height: number): pxt.sprite.ImageState {
     return {
-        bitmap: new Bitmap(width, height).data()
+        bitmap: new pxt.sprite.Bitmap(width, height).data()
     }
 }
 
-function cloneImage(state: ImageState): ImageState {
+function cloneImage(state: pxt.sprite.ImageState): pxt.sprite.ImageState {
     return {
         ...state,
-        bitmap: Bitmap.fromData(state.bitmap).copy().data(),
-        floatingLayer: state.floatingLayer && Bitmap.fromData(state.floatingLayer).copy().data()
+        bitmap: pxt.sprite.Bitmap.fromData(state.bitmap).copy().data(),
+        floatingLayer: state.floatingLayer && pxt.sprite.Bitmap.fromData(state.floatingLayer).copy().data()
     };
 }
